Move CORS headers to seperate function
