+++ functions/api/date.ts
@@ -1,3 +1,5 @@
+import {ApiHeaders} from "./headers";
+
 export async function onRequest(context) {
   // Contents of context object
   const {
@@ -8,38 +10,24 @@ export async function onRequest(context) {
     next, // used for middleware or to fetch assets
     data // arbitrary space for passing data between middlewares
   } = context;
-  // List of origins allowed to access API
-  // First origin will be returned unless another match is found
-  const allowedOrigins = ["http://time.tectic.ca", "http://localhost:8080"];
-  // Function to return CORS headers
-  const corsHeaders = (origin) => ({"Access-Control-Allow-Headers": "*", "Access-Control-Allow-Methods": "GET", "Access-Control-Allow-Origin": origin});
-  // Check origin. If allowed, return in header. Otherwise return first origin in list.
-  const checkOrigin = (request) => {
-    const foundOrigin = allowedOrigins.find((allowedOrigin) => allowedOrigin.includes(request.headers.get("Origin")));
-    return foundOrigin
-      ? foundOrigin
-      : allowedOrigins[0];
-  };
-  // Check if origin is valid
-  const allowedOrigin = checkOrigin(request);
+
   if (request.method === "OPTIONS") {
-    return new Response("OK", {headers: corsHeaders()});
-  }
-  if (request.method === "GET") {
-    // Application headers
-    const applicationHeaders = {
-      "content-type": "application/json;charset=UTF-8"
-    };
+    return new Response("OK", {headers: ApiHeaders(request)});
+  } else if (request.method === "GET") {
     // concat CORS and application
     const responseHeaders = {
-      ...corsHeaders(allowedOrigin),
-      ...applicationHeaders
+      ...ApiHeaders(request),
+      "content-type": "application/json;charset=UTF-8"
     };
     // create json response
     const dateJson = JSON.stringify({
       date: new Date(Date.now())
     });
     return new Response(dateJson, {headers: responseHeaders});
+  } else {
+    return new Response("Bad Request", {
+      status: 400,
+      headers: ApiHeaders(request)
+    });
   }
-  return;
 }
